Implemented AllInOne tests.
User can define a code file with the expected indentation, the test
method automatically checks if the calculated indentation is correct on
all file lines.
Implemented an event in the engine which is raised on the new line char.
On this event the user can check if the current line needs to be
reindented (thisLineIndent won't be changed afterwards, and moreover it
will be replaced with the next line indent).